refactor: use shared MultiModelSelector for allowed models
Replaces the admin dialog's hand-rolled AllowedModelsPicker with the
shared MultiModelSelector, fed with catalogue definitions filtered to
the models the catalog endpoint reports live on the provider. The
comma-separated paste support moves into the shared component so every
consumer gets it.
